improv: Refactor application code with a Layout type
