+++ .github/workflows/ci.yml
@@ -12,9 +12,9 @@ jobs:
   run-tests:
     runs-on: ubuntu-latest
     steps:
-    - run: sudo apt-get install cargo libgtk-3-dev libhidapi-dev libusb-1.0-0-dev xvfb
+    - run: sudo apt-get install cargo libgtk-3-dev libhidapi-dev libusb-1.0-0-dev
     - uses: actions/checkout@v2
-    - run: xvfb-run cargo test
+    - run: cargo test
 
   linux-x86_64:
     runs-on: ubuntu-latest
