+++ src/application/mod.rs
@@ -7,8 +7,8 @@ use crate::daemon::{Daemon, DaemonClient, daemon_server};
 
 mod key;
 mod keyboard;
+mod layout;
 mod page;
-mod physical_layout;
 mod picker;
 mod rect;
